+++ react/constants/Options.js
@@ -1,0 +1,7 @@
+export default {
+  LEFT: 'left',
+  MIDDLE: 'middle',
+  RIGHT: 'right',
+  INTERNAL: 'internal',
+  EXTERNAL: 'external'
+}